#172 Optimized the uploadImage() method closes #126
The method receives an image from a request, saves the image to a file, reads the image from the file, scales and crops the image, saves the image to a file again. That was a huge code smell. Now the image is only saved once after all the preocessing is done!