Change challenge complexity based on failed logins #243
Normally when a user loges in, he now has to type only 5 characters. This still leaves 916132832 possibilities. However, if someone tries to hack the server (requests many challenges) the length of the generated challenges increases to up to 10 characters (8.39E17 possibilities). Furthermore, all (and so also the 5 character long challenges) do no more work after 10 tries. So an attack on the first few short challenges does not work either.